requirements.txt
joblib==1.5.3
matplotlib==3.10.8
numba==0.63.1
numpy==2.3.5
pandas==3.0.0
scikit-learn==1.8.0
scipy==1.17.0
torch==2.9.1
torchvision==0.24.1
umap-learn==0.5.11
